Consolidated auto aim and aded autonomous commands (tested)
