+++ src/main/java/frc/robot/subsystems/drive/Drive.java
@@ -34,6 +34,7 @@ import edu.wpi.first.math.numbers.N1;
 import edu.wpi.first.math.numbers.N3;
 import edu.wpi.first.math.system.plant.DCMotor;
 import edu.wpi.first.math.util.Units;
+import edu.wpi.first.units.measure.Distance;
 import edu.wpi.first.wpilibj.Alert;
 import edu.wpi.first.wpilibj.Alert.AlertType;
 import edu.wpi.first.wpilibj.DriverStation;
@@ -44,6 +45,7 @@ import edu.wpi.first.wpilibj2.command.SubsystemBase;
 import edu.wpi.first.wpilibj2.command.sysid.SysIdRoutine;
 import frc.robot.Constants;
 import frc.robot.Constants.Mode;
+import frc.robot.Constants.PointsOfInterest;
 import frc.robot.generated.TunerConstants;
 import frc.robot.subsystems.auto.onTheFlySetpoints;
 import frc.robot.util.LocalADStarAK;
@@ -373,5 +375,57 @@ public class Drive extends SubsystemBase {
       getAutoAlignmentCommand(onTheFlySetpoints.CLIMB_LEFT_FINISH.blueAlignmentPose),
       getAutoAlignmentCommand(onTheFlySetpoints.CLIMB_RIGHT_SETUP.blueAlignmentPose),
       getAutoAlignmentCommand(onTheFlySetpoints.CLIMB_RIGHT_FINISH.blueAlignmentPose));
+  }
+
+  public Rotation2d findAngleForShooting(Pose2d pose) {
+    Distance positionX = pose.getMeasureX();
+    Distance positionY = pose.getMeasureY();
+
+    Distance deltaX;
+    Distance deltaY;
+
+    Rotation2d rotation;
+
+    if(isBlue()) {
+        if (positionX.minus(PointsOfInterest.centerOfHubBlue.getMeasureX()).in(Meters) > 0) { // If we are not in our zone
+            if (positionY.minus(PointsOfInterest.centerOfHubBlue.getMeasureY()).in(Meters) > 0) { // If we are in the north half of the field
+                deltaX = PointsOfInterest.cornerNW.getMeasureX().minus(positionX);
+                deltaY = PointsOfInterest.cornerNW.getMeasureY().minus(positionY);
+            } else { // If we are in the south half of the field
+                deltaX = PointsOfInterest.cornerSW.getMeasureX().minus(positionX);
+                deltaY = PointsOfInterest.cornerSW.getMeasureY().minus(positionY);
+            }
+
+            rotation = new Rotation2d(Math.atan(deltaY.baseUnitMagnitude() / deltaX.baseUnitMagnitude()) - Math.PI);
+        } else { // If we are in our zone (able to shoot at hub)
+            deltaX = PointsOfInterest.centerOfHubBlue.getMeasureX().minus(positionX);
+            deltaY = PointsOfInterest.centerOfHubBlue.getMeasureY().minus(positionY);
+
+            rotation = new Rotation2d(Math.atan(deltaY.baseUnitMagnitude() / deltaX.baseUnitMagnitude()));
+        }
+    } else {
+        if (positionX.minus(PointsOfInterest.centerOfHubRed.getMeasureX()).in(Meters) < 0) { // If we are not in our zone
+            if (positionY.minus(PointsOfInterest.centerOfHubRed.getMeasureY()).in(Meters) > 0) { // If we are in the north half of the field
+                deltaX = PointsOfInterest.cornerNE.getMeasureX().minus(positionX);
+                deltaY = PointsOfInterest.cornerNE.getMeasureY().minus(positionY);
+            } else { // If we are in the south half of the field
+                deltaX = PointsOfInterest.cornerSE.getMeasureX().minus(positionX);
+                deltaY = PointsOfInterest.cornerSE.getMeasureY().minus(positionY);
+            }
+
+            rotation = new Rotation2d(Math.atan(deltaY.baseUnitMagnitude() / deltaX.baseUnitMagnitude()));
+        } else { // If we are in our zone (able to shoot at hub)
+            deltaX = PointsOfInterest.centerOfHubRed.getMeasureX().minus(positionX);
+            deltaY = PointsOfInterest.centerOfHubRed.getMeasureY().minus(positionY);
+            
+            rotation = new Rotation2d(Math.atan(deltaY.baseUnitMagnitude() / deltaX.baseUnitMagnitude()) - Math.PI);
+        }
+    }
+        
+    return rotation;
+  }
+
+  public boolean isBlue() {
+      return DriverStation.getAlliance().orElse(Alliance.Red) == Alliance.Blue;
   }
 }
